refactor(lib): Rename storage-backend directory to store
